feat: Added a WordModel and WordRepository class for storage `word` information.
**WordRepository**
Class that us used to communicate with the Firestore words collection.

**WordModel**
Used internally by application and fetches and writes data to the Firestore service using the WordRepository class.

**SeedWordsUploader**
Read in /data/seed_words.csv information to a WordModel which gets saved to the Firestore `word` collection using the WordRepository class. This was done for proof that the classes created above would allow writing to the Firestore collection.

**Firestore Rules**
Added rules to the store to prevent writes unless your signed in to a Firebase account. Reads are open by default.