feat: Updated recorder to use Dart audio stream.
Needed a way to record audio to stream and allow a way to subscribe to that stream. This will be helpful for Speech-To-Text (STT) being able to transcribe the audio in real-time while the microphone is being recorded.

Added a decibel progress meter to allow for feedback from the microphone while recording.

Microphone permissions are check on start of the student word practice screen. Users will need to enable the microphone to being recording for the app.