feat: Created a `Cheetah` Speech-To-Text STT on-device assessor.
A https://picovoice.ai/docs/quick-start/cheetah-flutter/ was used to grab realtime audio from the Dart audio stream and create a transcript.

Interface for PronunciationAssessor was created as abstract class to create several on-device and cloud based Speech-To-Text (STT) providers that build out a transcript and compare that against the practice word for evaluating correctness. In this commit we just return the transcript and set up an example call to build out the AssessmentResult.