feat: Prevent students from accessing the teacher dashboard.
Should the student know their email they could enter it in the teacher login and access the teacher dashboard. Added a check after student login to see what role the user is. If the logged in account is a student role then they get bounced back to the user selection screen and if instructor role then they get access to the teacher dashboard.