Implementation of Profile Screen
Profile icon is shown now at the top of teacher dashboard, teacher student dashboard, and student dashboard screens. Features include current username, name, and email, along with a log out button to log the user out (no matter student or teacher) with an about link button which only directs users to the GitHub repo. Naturally, this link can be adapted to whatever help or about page we end up with later this week. ** Note: Opening the link may require some level of permission mutations on iOS, but it works on my Android testing device. Further note that the link may not work on emulators/simulators if there isn't a path to any browser to open the link in. Flutter also seems to get confused at times on hot restarts and then attempting to open the link because it doesn't always instantiate those browser dependencies on hot restarts.